+++ src/components/templates/ConversationSpace/ConversationSpace.tsx
@@ -23,8 +23,6 @@ import { RenderMarkdown } from "components/organisms/RenderMarkdown";
 import { Room } from "components/organisms/Room";
 
 import { InformationCard } from "components/molecules/InformationCard";
-import { TableComponent } from "components/molecules/TableComponent";
-import { TableGrid } from "components/molecules/TableGrid";
 import { TableHeader } from "components/molecules/TableHeader";
 import { TablesControlBar } from "components/molecules/TablesControlBar";
 import { UserList } from "components/molecules/UserList";
@@ -132,6 +130,8 @@ export const ConversationSpace: React.FC<ConversationSpaceProps> = ({
               )}
             </div>
           </div>
+          {/*
+            TODO
           <TableGrid
             setSeatedAtTable={setSeatedAtTable}
             seatedAtTable={seatedAtTable}
@@ -144,6 +144,8 @@ export const ConversationSpace: React.FC<ConversationSpaceProps> = ({
             venue={venue}
             template={VenueTemplate.conversationspace}
           />
+          */}
+
           <UserList
             usersSample={venue.recentUsersSample ?? ALWAYS_EMPTY_ARRAY}
             userCount={venue.recentUserCount ?? 0}
